Reload rules threadsafely
